achievment uncloked: get all errors

+++ lab1/src/algorithm.cpp
@@ -1,6 +1,6 @@
 #include "../include/algorithm.h"
 
-int doComputation(int a, double b) {
+int doComputation(int a, int b) {
   int c = max(3, 4);    
   logDebug('d');
   return c;
